Changed data type of amount from int to double

+++ Payment.java
@@ -1,22 +1,22 @@
 public class Payment
 {
-    private int amount;
+    private double amount;
     private Invoice invoice;
     private String datePaid;
     private String paymentType;
 
-    public Payment(int amount, Invoice invoice, String datePaid, String paymentType) {
+    public Payment(double amount, Invoice invoice, String datePaid, String paymentType) {
         this.amount = amount;
         this.invoice = invoice;
         this.datePaid = datePaid;
         this.paymentType = paymentType;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return this.amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
